Work in progress

+++ BookStore/Models/Category.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; }
 
-        public decimal Discount { get; set; }
+        public double Discount { get; set; }
     }
 }
